* Separation of library baseJc and simpleC: The simpleC-lib should contain only routines which runs under any standard-C and does not need any other libs. The OSAL.lib can use this simpleC-lib in its implementation. The order of lib should be well defined using GNU-linker.   * Definition of the C99-Standard types int32_t etc and use them in some of files. Reason: compatibility-problems with other sources in a special project. There was used 'int32' for a enum definition of an internal type. That is correct and should be accepted.   * New sources in Fwc for readline of files. But this source should not be a part of simpleC, because it uses the OSAL (TODO).   * New routines in fwc/simpleC.c to support working with strings. This simple routines are created respectively copied and adapted from older projects because the Standard-C scanf-routine is faulty on Visual Studio 6 and it is not really good. A user may not use the complex StringJc sources, the fwc-routines are lightweight.   * Some adaptions in InterProcessComm while using in a special project.   #Compilation checked in baseJc.lib, OSAL.lib, Reflectiontest.

+++ source/os_Windows/os_time.c
@@ -94,7 +94,7 @@ int32 os_getClockCnt(void)
 }
  
 
-void os_delayThread(ulong milliseconds)
+void os_delayThread(int32_t milliseconds)
 { Sleep(milliseconds);
 }
